LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/template-shadowrootmode-script-src.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <div><template shadowrootmode="closed"><script src="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=6f013501-95b1-4d24-abac-9e33c7c8ee6b&amp;encodingcheck=&Gbreve;"></script></template></div>


FAIL Speculative parsing, document.write(): template-shadowrootmode-script-src Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
